Move layer keys into much-easier to use spots
Only drawback is that the right half of the
keyboard is going to be harder to get into boot
mode from now on. Will require plugging in both
halves with right half plugged directly to
computer, using left half to enter boot mode, then
unplug the left half, then flash the right half.